Fix issue #8. Move flaki flatbuffers into api folder.
